refactor: migrate mobile app to riverpod feature architecture
