+++ mobile/lib/src/core/mobile_protocol.dart
@@ -1,0 +1,4 @@
+/// Version of the mobile companion protocol spoken with the Alera runtime
+/// gateway. The runtime enforces strict equality during `mobile.hello`, so new
+/// capabilities are feature-detected instead of bumping this value.
+const int aleraMobileProtocolVersion = 1;
