+++ mobile/lib/src/features/hosts/application/paired_hosts_controller.dart
@@ -1,0 +1,28 @@
+import 'package:alera_mobile/src/features/hosts/application/host_providers.dart';
+import 'package:alera_mobile/src/features/hosts/domain/paired_host_profile.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'paired_hosts_controller.g.dart';
+
+@riverpod
+class PairedHostsController extends _$PairedHostsController {
+  @override
+  Future<List<PairedHostProfile>> build() {
+    return ref.watch(hostRepositoryProvider).loadHosts();
+  }
+
+  Future<void> savePairedHost(
+    PairedHostProfile host,
+    String deviceToken,
+  ) async {
+    await ref.read(hostRepositoryProvider).savePairedHost(host, deviceToken);
+    ref.invalidateSelf();
+    await future;
+  }
+
+  Future<void> removeHost(String hostId) async {
+    await ref.read(hostRepositoryProvider).removeHost(hostId);
+    ref.invalidateSelf();
+    await future;
+  }
+}
