+++ mobile/lib/src/features/hosts/presentation/pair_host_screen.dart
@@ -1,21 +1,20 @@
+import 'package:alera_mobile/src/app/theme/alera_tokens.dart';
+import 'package:alera_mobile/src/features/hosts/application/paired_hosts_controller.dart';
+import 'package:alera_mobile/src/features/hosts/domain/paired_host_profile.dart';
+import 'package:alera_mobile/src/features/hosts/domain/pairing_offer.dart';
+import 'package:alera_mobile/src/features/runtime/infra/mobile_runtime_client.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:mobile_scanner/mobile_scanner.dart';
 
-import '../models.dart';
-import '../network/mobile_runtime_client.dart';
-import '../storage/host_repository.dart';
-import '../theme/alera_tokens.dart';
-
-class PairHostScreen extends StatefulWidget {
-  const PairHostScreen({super.key, required this.hostRepository});
-
-  final HostRepository hostRepository;
+class PairHostScreen extends ConsumerStatefulWidget {
+  const PairHostScreen({super.key});
 
   @override
-  State<PairHostScreen> createState() => _PairHostScreenState();
+  ConsumerState<PairHostScreen> createState() => _PairHostScreenState();
 }
 
-class _PairHostScreenState extends State<PairHostScreen> {
+class _PairHostScreenState extends ConsumerState<PairHostScreen> {
   final TextEditingController _controller = TextEditingController();
   bool _saving = false;
   bool _scannerOpen = false;
@@ -40,7 +39,9 @@ class _PairHostScreenState extends State<PairHostScreen> {
       final offer = PairingOffer.parse(rawOffer);
       final credentials = await MobileRuntimeClient.pairDevice(offer);
       final host = PairedHostProfile.fromPairingResult(offer, credentials);
-      await widget.hostRepository.savePairedHost(host, credentials.deviceToken);
+      await ref
+          .read(pairedHostsControllerProvider.notifier)
+          .savePairedHost(host, credentials.deviceToken);
       if (mounted) {
         Navigator.of(context).pop(true);
       }
